requirements.txt
Django>=5.1,<7.0
requests>=2.32,<3.0
gunicorn>=22.0,<24.0
psycopg[binary]>=3.2,<4.0
python-dotenv>=1.0,<2.0
